refactor(ai-chat): make the receipt row open review
Drop the per-row Review and Undo buttons: clicking the document opens its
changes, and undo already lives in the toolbar next to Done, where the reader
can see what they are undoing.
